Move functions into separate dir
